% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_thresholds.R
\name{check_thresholds}
\alias{check_thresholds}
\title{Check for overlapping thresholds}
\usage{
check_thresholds(
  data,
  low_col = low,
  medium_col = medium,
  high_col = high,
  very_high_col = very_high
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl} of thresholds with each pathogen/age
group assigned to a row and columns for low, medium, high, and very high
thresholds.}

\item{low_col}{Name of column showing low threshold (unquoted)}

\item{medium_col}{Name of column showing medium threshold (unquoted)}

\item{high_col}{Name of column showing high threshold (unquoted)}

\item{very_high_col}{Name of column showing very high threshold (unquoted)}
}
\value{
If overlapping thresholds detected then \code{tbl} is returned with
flagged thresholds
}
\description{
Check for overlapping thresholds
}
\examples{

x <- dplyr::tibble(
  indicator = c("a", "b"),
  low = c(0.5, 1.2),
  medium = c(1.1, 0.8),
  high = c(2.5, 3.1),
  very_high = c(3.7, 4.2)
  )

check_thresholds(x)
}
