% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats.R
\name{epi_stats}
\alias{epi_stats}
\title{Calculate epidemiological stats}
\usage{
epi_stats(data, season, epidemic_threshold, week_col = iso_week)
}
\arguments{
\item{data}{Aggregated weekly counts as \code{tbl} or \code{data.frame} with each week
as a row and a column for rate}

\item{season}{Character string ("YYYY/YYYY") of season to be checked against (see \link[threshtools]{find_flu_season})}

\item{epidemic_threshold}{Numeric value representing low (epidemic) threshold}

\item{week_col}{Name of column with week number (default = iso_week)}
}
\value{
Named vector with first week above threshold (\code{epi_start}) and number of weeks above threshold (\code{epi_length})
}
\description{
Function that calculates the first week of a chosen season
above a defined epidemic threshold and the total number of weeks that season
remains above threshold
}
\examples{

x <- tibble::tribble(
 ~week, ~year, ~flu_season,  ~rate,
 40,    2023,  "2023/2024",  2.5,
 41,    2023,  "2023/2024",  3.1,
 42,    2023,  "2023/2024",  4.2,
 40,    2024,  "2024/2025",  2.5,
 41,    2024,  "2024/2025",  3.1,
 42,    2024,  "2024/2025",  4.2,
 40,    2025,  "2025/2026",  2.5,
 41,    2025,  "2025/2026",  3.1,
 42,    2025,  "2025/2026",  4.2,
 )

 epi_stats(
  x,
  season = "2024/2025",
  epidemic_threshold = 2.9,
  week_col = week
 )
}
