% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series}
\usage{
plot_ts(
  data,
  x = week_date,
  y = rate,
  geom = "line",
  fill = NULL,
  colour = NULL,
  group = NULL
)
}
\arguments{
\item{data}{\code{tbl} or \code{data.frame} of aggregated weekly rates with weeks as rows
and columns with week as class \code{date} (hint: use as.Date(grates::isoweek())
to get date from week and year columns)}

\item{x}{Column to be ploted on x axis (default = week_date)}

\item{y}{Column to be plotted on y axis (default = rate)}

\item{geom}{A character string ("line" or "col") describing whether line graph or epi curve should be plotted}

\item{fill}{Group to be differentiated by fill}

\item{colour}{Group to be differentiated by colour}

\item{group}{Grouping to be provided to ggplot2}
}
\value{
ggplot object with rates over time
}
\description{
Take aggregate data and plot basic time-series using \link[ggplot2]{ggplot}
}
\examples{

x <- tibble::tribble(
~week, ~year, ~flu_season,  ~rate,
40,    2023,  "2023/2024",  2.5,
41,    2023,  "2023/2024",  3.1,
42,    2023,  "2023/2024",  4.2,
40,    2024,  "2024/2025",  2.5,
41,    2024,  "2024/2025",  3.1,
42,    2024,  "2024/2025",  4.2,
40,    2025,  "2025/2026",  2.5,
41,    2025,  "2025/2026",  3.1,
42,    2025,  "2025/2026",  4.2,
)

plot_ts(x)
}
