% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_msd.R
\name{calculate_msd}
\alias{calculate_msd}
\title{Calculate MSD thresholds}
\usage{
calculate_msd(
  data,
  rate_col = rate,
  season_col = flu_season,
  decimal_places = 2
)
}
\arguments{
\item{data}{Aggregated weekly rates of class \code{tbl} or \code{data.frame} with weeks as rows
and a column for rates}

\item{rate_col}{Column containing rates (default = rate)}

\item{season_col}{Column containing flu season (see \link[threshtools]{find_flu_season})}

\item{decimal_places}{Number of decimal places to round thresholds}
}
\value{
\code{tbl} with columns for mean rate, standard deviation of rate,
number of seasons included, low, medium, high and very high threshold
}
\description{
Functions calculate activity level thresholds from aggregated weekly rates
using the mean standard deviation (MSD) method \url{https://doi.org/10.2807/1560-7917.ES.2024.29.45.2400696}\link{(Sinnathamby et al. 2024)}
}
\examples{

#' # Create dummy data
x <- tibble::tribble(
~week, ~year, ~flu_season,  ~rate,
40,    2023,  "2023/2024",  2.5,
41,    2023,  "2023/2024",  3.1,
42,    2023,  "2023/2024",  4.2,
40,    2024,  "2024/2025",  2.5,
41,    2024,  "2024/2025",  3.1,
42,    2024,  "2024/2025",  4.2,
40,    2025,  "2025/2026",  2.5,
41,    2025,  "2025/2026",  3.1,
42,    2025,  "2025/2026",  4.2,
)

calculate_msd(x)
}
