% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_mem.R
\name{data_to_mem}
\alias{data_to_mem}
\title{Transform aggregated data table into format accepted by \code{mem::memmodel()}}
\usage{
data_to_mem(
  data,
  seasons,
  duplicate = TRUE,
  season_length = 33,
  year_col = year,
  week_col = iso_week,
  season_col = flu_season,
  rate_col = rate
)
}
\arguments{
\item{data}{Aggregated weekly counts of class \code{data.table} or \code{tbl} of with
rows as weeks columns for rate and flu season
(see \code{threshtools::find_flu_season()})}

\item{seasons}{A vector or single character string in format \emph{"YYYY/YYYY"}}

\item{duplicate}{Boolean operator which controls whether single seasons
should be duplicated to force acceptance by \code{mem::model()} (default = \code{TRUE})}

\item{season_length}{Number of weeks in flu season (default = 33)}

\item{year_col}{Column with year}

\item{week_col}{Column with week}

\item{season_col}{Column with flu season (see \link[threshtools]{find_flu_season})}

\item{rate_col}{Column with rate}
}
\value{
Transformed \code{tbl} with rows as weeks and each season as a column.
Week 53 is preserved if present as it is handled by \code{mem::model()} and
NAs are supplied to seasons without week 53.
}
\description{
Transform aggregated data table into format accepted by \code{mem::memmodel()}
}
\examples{

# Create dummy data
x <- tibble::tribble(
~week, ~year, ~flu_season,  ~rate,
40,    2023,  "2023/2024",  2.5,
41,    2023,  "2023/2024",  3.1,
42,    2023,  "2023/2024",  4.2,
40,    2024,  "2024/2025",  2.5,
41,    2024,  "2024/2025",  3.1,
42,    2024,  "2024/2025",  4.2,
40,    2025,  "2025/2026",  2.5,
41,    2025,  "2025/2026",  3.1,
42,    2025,  "2025/2026",  4.2,
)
data_to_mem(
x,
seasons = c("2023/2024", "2024/2025"),
week_col = week,
year_col = year,
season_col = flu_season,
rate_col = rate,
season_length = 33
)

}
