% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mem.R
\name{tidy_mem}
\alias{tidy_mem}
\title{Tidy \code{memmodel} outputs}
\usage{
tidy_mem(mem_model, decimal_places = 2, ecdc_format = FALSE)
}
\arguments{
\item{mem_model}{Output from \code{mem::memmodel()} (see \link[mem]{memmodel})}

\item{decimal_places}{Number of decimal places thresholds should be rounded to}

\item{ecdc_format}{A boolean operator (default = \code{FALSE}). If \code{TRUE} uses
format required for annual submission to ECDC which includes post-epidemic values.}
}
\value{
\code{tbl} of low, medium, high and very high thresholds, with key stats on
mean start week, mean length of epidemic, effective number of pre-epidemic
values (\code{n_max}) and number of seasons included in the model (\code{n_seasons})
}
\description{
Extracts \code{mem::memmodel()} outputs and transforms into tidy format
}
